factorized the addGif (used now in search as well), created the test suite
