+++ src/search.ts
@@ -86,7 +86,7 @@ export const search: (
 	editor: vscode.TextEditor,
 	context: vscode.ExtensionContext,
 	history: HistoryProvider
-) => Promise<boolean> = async (
+) => Promise<boolean | undefined> = async (
 	editor: vscode.TextEditor,
 	context: vscode.ExtensionContext,
 	history
@@ -111,14 +111,12 @@ export const searchTask: (
 	editor: vscode.TextEditor,
 	context: vscode.ExtensionContext,
 	history: HistoryProvider
-) => Promise<boolean> = async (
+) => Promise<boolean | undefined> = async (
 	searchInput: string | undefined,
 	editor: vscode.TextEditor,
 	context: vscode.ExtensionContext,
 	history: HistoryProvider
 ) => {
-	// grabbing the current location to insert the edit later with the GIFLENS tag
-	const position: vscode.Position = editor.selection.active;
 	if (searchInput) {
 		try {
 			// part about getting the data and creating the img html tags for the images.
@@ -136,7 +134,11 @@ export const searchTask: (
 				getChosenGifUrl(searchResults, resolve, searchInput, context, history)
 			);
 
-			return addGifLensTagToEditor(editor, position, urlToUse);
+			return vscode.commands.executeCommand(
+				'giflens.addGif',
+				urlToUse.url,
+				editor
+			);
 		} catch (err) {
 			// if an error is returned from the search, calls the handleAPIError function
 			return handleApiError(err);
@@ -149,54 +151,6 @@ export const searchTask: (
 		);
 		return false;
 	}
-};
-
-/**
- * add a giflens tag with the specified url in the active editor
- * @param  {vscode.TextEditor} editor the active vscode editor
- * @param  {vscode.Position} position the position of the cursor in the active vscode editor
- * @param  {Gif} gif the url of the gif for which to insert a GifLens Tag
- * @returns {Thenable<boolean} a final status of the insertion
- */
-const addGifLensTagToEditor: (
-	editor: vscode.TextEditor,
-	position: vscode.Position,
-	url: Gif
-) => Thenable<boolean> = (editor, position, gif) => {
-	return editor.edit(editBuilder => {
-		// getting the position where to insert (beginning of the current line)
-		let positionToInsert = new vscode.Position(position.line, 0);
-		// first case when the selected line is empty, we do not create a new line
-		if (editor.document.lineAt(position).isEmptyOrWhitespace) {
-			editBuilder.insert(
-				positionToInsert,
-				`${getLanguageCommentStart(editor.document.languageId)} GIFLENS-${
-					gif.url
-				}${getLanguageCommentEnd(editor.document.languageId)}`
-				// \r is used to create a new line, VSCode converts automatically to the end of line of the current OS
-			);
-			// else second case when using it from a line of code, we insert a new line above
-		} else {
-			// getting the number of spaces or tabs at the beginning of the line
-			const lineBeginningChars: number = editor.document.lineAt(position)
-				.firstNonWhitespaceCharacterIndex;
-			// goes to the beginning of the line to create the GIFLENS tag the line above after insertion
-			editBuilder.insert(
-				positionToInsert,
-				`${
-					// insertSpaces returns false if the user uses tabs, true if the user uses spaces
-					// it is defined per document in VSCode, so if the user voluntarily changes it on one line, this code will not work
-					editor.options.insertSpaces
-						? // returns the correct indentation character for the user
-						  ' '.repeat(lineBeginningChars)
-						: '\t'.repeat(lineBeginningChars)
-				}${getLanguageCommentStart(editor.document.languageId)} GIFLENS-${
-					gif.url
-				}${getLanguageCommentEnd(editor.document.languageId)}\r`
-				// \r is used to create a new line, VSCode converts automatically to the end of line of the current OS
-			);
-		}
-	});
 };
 
 /**
@@ -297,83 +251,6 @@ export const createGifSelectionPanel: (
 			? webviewHtml(images)
 			: webviewHtml(images, page, true);
 	return panel;
-};
-
-/**
- * The syntax to open a comment for a specific language.
- * @param languageId The languageId (handled by VS code).
- * @returns A String to open a comment.
- */
-export const getLanguageCommentStart = (languageId: String) => {
-	switch (languageId) {
-		case 'bat':
-			return 'REM';
-		case 'clojure':
-			return ';';
-		case 'ruby':
-		case 'coffeescript':
-		case 'dockerfile':
-		case 'makefile':
-		case 'perl':
-		case 'powershell':
-		case 'python':
-		case 'r':
-		case 'shellscript':
-		case 'yaml':
-			return '#';
-		case 'c':
-		case 'css':
-			return '/*';
-		case 'html':
-		case 'markdown':
-			return '<!--';
-		case 'lua':
-		case 'sql':
-			return '--';
-		case 'swift':
-			return '///';
-		case 'vb':
-			return "'";
-		case 'javascript':
-		case 'typescript':
-		case 'cpp':
-		case 'csharp':
-		case 'fsharp':
-		case 'go':
-		case 'groovy':
-		case 'java':
-		case 'javascriptreact':
-		case 'less':
-		case 'objective-c':
-		case 'objective-cpp':
-		case 'php':
-		case 'jade':
-		case 'rust':
-		case 'scss':
-		case 'sass':
-		case 'typescriptreact':
-		default:
-			return '//';
-	}
-};
-
-/**
- * The (optional) closing comment syntax for the language.
- * All the results need to start with a space!
- * @param languageId The languageId (handled by VS code).
- * @returns A String, mepty for most cases.
- */
-export const getLanguageCommentEnd = (languageId: String) => {
-	switch (languageId) {
-		case 'c':
-		case 'css':
-			return ' */';
-		case 'html':
-		case 'markdown':
-			return ' -->';
-		default:
-			return '';
-	}
 };
 
 /**
